Generalize and move queue types to sophon mod root
